# ruby: Improve performance of Google::Protobuf::Timestamp#to_time